backend/requirements.txt
flask
kafka-python
flask-cors
psycopg2-binary